pyproject.toml
[project]
name = "accommodation-buddy"
version = "0.1.0"
description = "Web application that helps teachers accommodate lesson materials for multilingual learners"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "celery[redis]>=5.4",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
    "python-docx>=1.1",
    "python-pptx>=1.0",
    "pdf2image>=1.17",
    "Pillow>=10.0",
    "wordfreq>=3.1",
    "python-magic>=0.4",
    "passlib[bcrypt]>=1.7",
    "pydantic-settings>=2.0",
    "itsdangerous>=2.1",
    "psycopg2-binary>=2.9.11",
]

[project.scripts]
accommodation-buddy = "accommodation_buddy.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/accommodation_buddy"]
